Making fixex so it works better for macs

+++ scripts/python.sh
@@ -1,6 +1,13 @@
 #!/usr/bin/env bash
 
+if [ "$OS" = "darwin" ]; then
+  sudo easy_install pip
+fi
+
 pip install --upgrade pip
 
 pip install --user neovim
-pip3 install --user neovim
+
+if command_exists pip3; then
+  pip3 install --user neovim
+fi
